Rename container files
