+++ client/src/components/AppDetailsContainer.scss
